refactor: unify rankings hook and clean duplicates
